.登录优化

+++ home-api/src/main/java/com/mlk/home/service/ManagerBaseService.java
@@ -9,7 +9,7 @@ import java.util.List;
  * Created by malikai on 2018-5-22.
  */
 public interface ManagerBaseService {
-    Long register(ManagerLogin model);
+    Integer register(ManagerLogin model);
 
     Boolean binding(ManagerFamilyGroup group);
